Make script execution faster
It was very inefficient on native platforms, with native tests taking up to 1h on ubuntu and osx64 instead of 30 seconds on the JVM.
The root cause was that native kotlin collections are very inefficient and some of the patterns we imported from Scala do not work.
For example calling tail() on immutable lists is very inefficient (which is suprising ciming from Scala).
Here we just use mutable collections in the script execution loop, which is not recursive anymore.