Simple changes to make this library more scala-friendly
This will make bitcoin-kmp easier to use from bitcoin-lib and eclair.
The biggest change is the definition of TaprootTweak, which has now a flat hierarchy (it was just unusable from
Scala and I did not want to create a Scala clone for this flag).

We also upgrade secp256k1-kmp which now is based on secp256k1 0.2.0