Introduce `TxId` and `TxHash` types (#90)
It was really footgunny to only use a `ByteVector32` here, as users could
use a `TxHash` when a `TxId` was expected. Introducing explicit types
ensures that functions can clearly specify which one they expect.

We also introduce a `BlockId` and `BlockHash` type.
It was footgunny to use a `ByteVector32` here as well, as we can easily
mistake one for the other.