Refactor Taproot ScriptTree and ControlBlock
We let the caller build arbitrary script trees. We then provide helpers
that build the merkle proofs and control blocks based on the leaf they
would like to spend.

It better encapsulate the details of how to build the control block and
the merkle proofs, which callers shouldn't need to worry about.