Provide high-level taproot abstractions (#109)
* Refactor Taproot ScriptTree and ControlBlock

We let the caller build arbitrary script trees. We then provide helpers
that build the merkle proofs and control blocks based on the leaf they
would like to spend.

It better encapsulate the details of how to build the control block and
the merkle proofs, which callers shouldn't need to worry about.

* Add helpers to spend taproot key path or script paths

We provide helpers for spending taproot output via the key path or any
script path, without dealing with low-level details such as signature
version, control blocks or script execution context.

It makes it easier and less error-prone to spend taproot outputs in
higher level applications.