Refactor and simplify how private/public keys are tweaked for taproot spends (#170)
* Refactor and simplify how private/public keys are tweaked for taproot spends

- We should not expose a Script.pay2tr() methods that sends to a raw untweaked key. It's confusing and users should never need it anyway.
- SchnorrTweak.NoTweak was not needed, it is simpler to just use an optional tweak when signing with the private key
-Remove default constructor for TaprootTweak

Users now have to explicitly create a KeyPath or ScriptPath tweak.

* Use secp256k1-kmp 0.22.0

* Don't use nullable script tree/script root parameters in Script.pay2tr()

Users had to cast null to ByteVector32?/ScriptTree? which was ugly. Now they
have to pass in either an explicit script/script root or use an explicit TaprootTweak.

* Remove google repository and sonatype snasphot repository

We can always temporarly add the snapshot repo back in PRs that test upstream snasphot dependencies (secpc256k1-kmp) and remove it when we update to a released version.